Improved: renamed the rule creation pages (#238)
