+++ src/views/CreateThresholdRule.vue
